final_sensen_security/sensen_security/requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.0.1
pandas==2.1.3
plotly==5.17.0
django-extensions==3.2.3